requirements.txt
# ============================================================
# 核心數值計算 | Core numerical computing
# ============================================================
numpy>=1.21.0
scipy>=1.7.0

# ============================================================
# 資料處理與分析 | Data processing and analysis
# ============================================================
pandas>=1.3.0

# ============================================================
# Excel 檔案處理 | Excel file processing
# ============================================================
openpyxl>=3.0.9

# ============================================================
# 圖表與視覺化 | Plotting and visualization
# ============================================================
matplotlib>=3.4.0

# ============================================================
# Web 框架與伺服器 | Web framework and server
# ============================================================
Flask>=3.0.0
Flask-CORS>=4.0.0
gunicorn>=21.0.0

# ============================================================
# 環境變數管理 | Environment variable management
# ============================================================
python-dotenv>=1.0.0

# ============================================================
# 類型提示支援 | Type hints support (Python < 3.9)
# ============================================================
typing-extensions>=4.0.0

# ============================================================
# 開發與測試工具 | Development and testing tools
# ============================================================
pytest>=6.0.0
